pep compliant and move the environment creation (merge of facade info and worker env) to the general worker.
