+++ workers/worker.py
@@ -1,6 +1,8 @@
 from abc import ABCMeta
 import logging
 from queue_managers import get_message_subject, send_message, subscribe
+import random
+import string
 from workers.common.types_management import start_publishing_type, \
     stop_publishing_type
 
@@ -48,3 +50,38 @@ class Worker(object):
             callback(self, message)
         else:
             logging.error('No callback for %s found!', subject)
+
+    def _create_container_environment(self, message):
+        """
+        Merge the build_parameters configured for the worker and gotten from
+        the facade.
+        :rtype : list
+        Policy: Only override worker-local parameters. If a parameter is empty,
+        generate a value using lowercase, uppercase and digits.
+        :param message: dict containing the message gotten from the bus
+        :return: list containing key=value pairs send to docker
+        """
+        if 'environment' in message.keys():
+            injected_parameters = message['environment']
+        else:
+            injected_parameters = dict()
+        if 'environment' in self.worker_info.keys():
+            local_parameters = self.worker_info['environment']
+        else:
+            local_parameters = dict()
+        environment = []
+        for key in local_parameters.keys():
+            if key in injected_parameters.keys():
+                environment.append(key + '=' + injected_parameters[key])
+            else:
+                if local_parameters[key] == '':
+                    environment.append(key + '=' + ''.join(
+                        random.SystemRandom().choice(
+                            string.ascii_lowercase +
+                            string.ascii_uppercase +
+                            string.digits)
+                        for _ in range(16)))
+                else:
+                    environment.append(key + '=' + local_parameters[key])
+        logging.debug('Current environment for VM: %s', environment)
+        return environment
